🔀 refactor(InitMediaCommand.php): rename namespace and class from MakeComponentCommand to InitMediaCommand for better clarity and consistency
🔀 refactor(StarterKitServiceProvider.php): add InitMediaCommand to the list of commands in the service provider to enable the command in the application
